Small cleanup of base package
